requirements.txt
gradio>=5.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=1.0.0
pypdf>=5.0.0
pdf2image>=1.17.0
reportlab>=4.0.0
docx2txt>=0.8
pillow>=10.0.0

